requirements.txt
mysqlclient==2.0.3
mysql-connector-python==8.0.23
pandas>=0.19
sqlalchemy==1.3.19
xlrd==1.2.0
streamlit==1.0.0
plotly==4.14.3
yagmail==0.14.256